refactor: 홈 화면 컴포넌트 추가 분리 및 불필요한 css 파일 제거
