+++ packages/service/src/page/main/index.tsx
@@ -1,23 +1,12 @@
-import { Suspense } from 'react';
-
-import { MainContent } from '@/widget/main/content';
-import { MainContentLoadingFallback } from '@/widget/main/content/loading';
-import { MainFirstSection } from '@/widget/main/section/first';
+import { MainBody } from '@/widget/main/body';
 
 import { MainFooter } from '@moeasy/storybook/ui/footer';
 import { Separator } from '@moeasy/storybook/ui/separator';
 
-import * as styles from './main.css';
-
 export async function MainPage() {
   return (
     <>
-      <main className={styles.main}>
-        <MainFirstSection />
-        <Suspense fallback={<MainContentLoadingFallback />}>
-          <MainContent />
-        </Suspense>
-      </main>
+      <MainBody />
       <Separator direction="horizontal" />
       <MainFooter />
     </>
