--- packages/service/src/page/main/main.css.ts
@@ -1,10 +1,0 @@
-import { headerHeight } from '@moeasy/storybook/utils/styles/global.css';
-
-import { style } from '@vanilla-extract/css';
-
-export const main = style({
-  display: 'flex',
-  flexDirection: 'column',
-  alignItems: 'center',
-  minHeight: `calc(100vh - ${headerHeight})`,
-});
